requirements.txt
flask
selenium
flask-restful
Pillow
webdriver-manager
Gunicorn
gevent
qrcode